+++ docs/apidoc.md
@@ -1,4 +1,4 @@
-::: sweet_validation.models
+::: sweet_validation.columns
     options:
       heading_level: 3
