--- sweet_validation/models/__init__.py
@@ -1,1 +1,0 @@
-from .columns import *  # noqa
