+++ sweet_validation/tests/test_columns.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 
-from sweet_validation.models import BaseColumn
-from sweet_validation.models.exceptions import DuplicatedValueError, RequiredValueError
+from sweet_validation.columns import BaseColumn
+from sweet_validation.exceptions import DuplicatedValueError, RequiredValueError
 
 
 def test_column_create():
